+++ include/observation_parser.hh
@@ -6,11 +6,11 @@
 
 namespace kalman {
     struct obs_elem {
-    double position;
-    double thickness;
-    double luminosity;
+    float position;
+    float thickness;
+    float luminosity;
     obs_elem() = default;
-    obs_elem(double a, double b, double c)
+    obs_elem(float a, float b, float c)
     : position(a), thickness(b), luminosity(c)
     {}
 };
@@ -19,7 +19,7 @@ namespace kalman {
         static std::pair<obs_elem*, unsigned int*> parse_gpu(int width, int height, std::vector<u_int8_t> &img, int threshold);
         // static std::pair<obs_elem*, unsigned int*> parse_gpu2(int width, int height, std::vector<u_int8_t> &img, std::ptrdiff_t stride, int threshold);
         // static std::pair<obs_elem*, unsigned int*> parse_gpu3(int width, int height, std::vector<u_int8_t> &img, int threshold);
-        std::vector<std::vector<kMatrix<double>>> parse(int width, int height, std::vector<u_int8_t> img, int threshold);
+        std::vector<std::vector<kMatrix<float>>> parse(int width, int height, std::vector<u_int8_t> img, int threshold);
     };
 
     void test(void);
